requirements.txt
flask
flask_sqlalchemy
flask_login
python-dotenv
google-generativeai
werkzeug
flask-migrate